Moved cloud connection serializer inside the model
