+++ src/backend/api/models/__init__.py
@@ -1,3 +1,3 @@
 from .invalid_token import InvalidToken
 from .copy_job import CopyJob, CopyJobSerializer
-from .cloud_connection import CloudConnection
+from .cloud_connection import CloudConnection, CloudConnectionSerializer
