+++ src/backend/api/models/cloud_connection.py
@@ -1,5 +1,7 @@
 import datetime
 
+from flask_restplus import Namespace, fields
+
 from ..application import db
 
 
@@ -22,3 +24,19 @@ class CloudConnection(db.Model):
 
     def __repr__(self):
         return "<Cloud Connection {}>".format(self.name)
+
+
+
+class CloudConnectionSerializer:
+    api = Namespace('connections', description='Connection related operations')
+    dto = api.model('connection', {
+        'id': fields.Integer(readonly=True, example=1),
+        'type': fields.String(required=True, example='S3'),
+        'name': fields.String(required=True, example='arbitrary-unique-name'),
+        'bucket': fields.String(required=True, example='my-bucket-name'),
+        'region': fields.String(required=True, example='us-west-2'),
+        'access_key_id': fields.String(required=True, example='KJRHJKHWEIUJDSJKDC2J'),
+        'access_key_secret': fields.String(required=True, example='jksldASDLASdak+asdSDASDKjasldkjadASDAasd'),
+        # access_key examples above have the correct length, but characters are made up
+    })
+
